django_project/requirements.txt
amazon.ion==0.5.0
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jwt==0.6.1
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mercurial==5.3
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycparser==2.19
pygame==1.9.6
Pygments==2.5.2
pylint==2.4.4
PyNaCl==1.3.0
pyrsistent==0.15.6
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
redshift==0.0.1
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.7
virtualenv==16.7.9
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
